Splitter en fil for å tilfredsstille sonar
